Add reusable slider component
